pyproject.toml
[project]
name = "ruoyi-py"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.2.0",
    "fastapi>=0.119.0",
    "pydantic>=2.12.0",
    "pydantic-yaml>=1.6.0",
    "redis>=6.4.0",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.37.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
package = false
